requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
dbt-core==1.8.4
dbt-postgres==1.8.2
psycopg2-binary==2.9.9
requests==2.32.3
pandas





